+++ kiltergpt/utils/plotter.py
@@ -25,7 +25,7 @@ class Plotter:
     def _create_image_coords(self, image_coords: pd.DataFrame):
         return {name: (row["img_x"], row["img_y"]) for name, row in image_coords.iterrows()}
 
-    def plot_climb(self, frames: str, return_fig: bool = False, highlight: str = None):
+    def plot_climb(self, frames: str, return_fig: bool = False, highlight: str | None = None):
         assert all(x in "0123456789pr" for x in frames), "Frames should only contain p, r and digits"
         frames = frames.replace(" ", "")  # here the input takes no whitespace
         board_path = "figs/full_board_commercial.png"
@@ -60,5 +60,5 @@ class Plotter:
             return plt.imshow(image)
         return image
 
-    def __call__(self, frames: str, return_fig: bool = False, highlight: str = None):
+    def __call__(self, frames: str, return_fig: bool = False, highlight: str | None = None):
         return self.plot_climb(frames, return_fig, highlight)